requirements.txt
Pillow
pypdf
pypdf[image]
